refactor(e2e): use a single pixOrgaPage
