--- high-level-tests/e2e-playwright/pages/pix-orga/LoginPage.ts
@@ -1,16 +1,0 @@
-import { Page } from '@playwright/test';
-
-export class LoginPage {
-  constructor(private readonly page: Page) {}
-
-  async login(email: string, rawPassword: string) {
-    await this.page.getByLabel('Adresse e-mail').fill(email);
-    await this.page.getByLabel('Mot de passe').fill(rawPassword);
-
-    await this.page.getByRole('button', { name: 'Je me connecte' }).click();
-  }
-
-  async acceptCGU() {
-    await this.page.getByRole('button', { name: 'Accepter et continuer' }).click();
-  }
-}
