+++ high-level-tests/e2e-playwright/pages/pix-orga/index.ts
@@ -1,3 +1,1 @@
-export * from './CreateCampaignPage.ts';
-export * from './LoginPage.ts';
 export * from './PixOrgaPage.ts';
